Create relay base class.
