+++ piksi_multi_cpp/CMakeLists.txt
@@ -21,7 +21,8 @@ cs_add_library(${PROJECT_NAME}
   src/receiver/receiver.cc
   src/sbp_callback_handler/sbp_callback_handler.cc
   src/sbp_callback_handler/sbp_callback_handler_factory.cc
-  src/sbp_callback_handler/sbp_callback_handler_heartbeat.cc
+  src/sbp_callback_handler/sbp_callback_handler_relay.cc
+  src/sbp_callback_handler/sbp_callback_handler_relay_heartbeat.cc
 )
 target_link_libraries(${PROJECT_NAME} serialport)
