+++ piksi_multi_cpp/include/piksi_multi_cpp/sbp_callback_handler/sbp_callback_handler_factory.h
@@ -12,9 +12,10 @@ factories. */
 class SBPCallbackHandlerFactory {
  public:
   // Factory method to create a callback that relays a specified sbp_msg_type.
-  static SBPCallbackHandler::SBPCallbackHandlerPtr createSBPRelayCallbackBySBPMsgType(
-      const ros::NodeHandle& nh, const uint16_t sbp_msg_type,
-      const std::shared_ptr<sbp_state_t>& state);
+  static SBPCallbackHandler::SBPCallbackHandlerPtr
+  createSBPRelayCallbackBySBPMsgType(const ros::NodeHandle& nh,
+                                     const uint16_t sbp_msg_type,
+                                     const std::shared_ptr<sbp_state_t>& state);
 };
 }  // namespace piksi_multi_cpp
